requirements.txt
playwright==1.44.0
selenium==4.21.0
schedule==1.2.2
pyyaml==6.0.1
requests==2.32.3
pandas==2.2.2
python-dotenv==1.0.1
pytest==8.2.2
pytest-asyncio==0.23.7
